Move actions to core folder
