+++ app/core/actions/series.js
