+++ app/core/sagas/index.js
@@ -13,11 +13,11 @@ import { remote } from 'electron';
 
 import Events from '../../events';
 import Api from './api';
-import { series } from '../../actions/series';
-import { api } from '../../actions/api';
-import { creators as orm } from '../../actions/orm';
-import uiActions from '../../actions/ui';
-import dashboardActions from '../../actions/dashboard';
+import { series } from '../../core/actions/series';
+import { api } from '../../core/actions/api';
+import { creators as orm } from '../../core/actions/orm';
+import uiActions from '../../core/actions/ui';
+import dashboardActions from '../../core/actions/dashboard';
 import watchRequests, { queueRequest } from './ApiRequestQueue';
 import notificationsWatcher, {
   queueRequest as queueNotification
