+++ app/reducers/series.js
@@ -1,5 +1,5 @@
 import { handleAction } from 'redux-actions';
-import { SERIES } from '../actions/series';
+import { SERIES } from '../core/actions/series';
 
 const series = handleAction(
   SERIES,
